pyproject.toml
[tool.poetry]
authors = ["David Cooper <david@dtcooper.com>"]
description = "Libraty providing a buffer interface to your Raspberry Pi's GPU layer. Usable with pygame, PIL and other graphics libraries."
homepage = "https://github.com/dtcooper/python-dispmanx"
include = ["LICENSE"]
license = "MIT"
name = "dispmanx"
readme = "README.md"
repository = "https://github.com/dtcooper/python-dispmanx"
version = "0.0.1-dev4"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.9",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: pygame",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.9"
numpy = {version = "<2", optional = true}

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
mkdocs = "^1.3.1"
mkdocs-autorefs = "^0.4.1"
mkdocs-material = "^8.3.9"
mkdocs-section-index = "^0.3.4"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}
mypy = "^0.971"
Pillow = "^9.2.0"
pycairo = "^1.21.0"
pygame = "^2.1.2"

[tool.poetry.extras]
numpy = ["numpy"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
preview = true

[tool.isort]
multi_line_output = 3
lines_after_imports = 2
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
force_sort_within_sections = true
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'GRAPHICSTHIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']
known_graphicsthirdparty = ['pygame', 'PIL']
